phase-2/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]==0.115.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic-settings==2.4.0",
    "python-dateutil==2.9.0.post0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "requests==2.32.3",
    "sqlalchemy-json==0.7.0",
    "sqlmodel==0.0.22",
    "uvicorn==0.30.0",
]
